[ISSUE #9657]♻️Extract Tokio client connect-flight orchestration
