+++ rocketmq-transport/src/clients/rocketmq_tokio_client.rs
@@ -39,7 +39,6 @@ use tracing::warn;
 
 use crate::base::connection_net_event::ConnectionNetEvent;
 use crate::base::pending_request_table::PendingRequestTable;
-use crate::clients::client::SessionConnectTarget;
 use crate::clients::nameserver_endpoint::ConnectTarget;
 use crate::clients::nameserver_endpoint::NameServerEndpoint;
 use crate::clients::TransportSession;
@@ -65,6 +64,7 @@ use rocketmq_protocol::code::response_code::ResponseCode;
 use rocketmq_protocol::protocol::remoting_command::RemotingCommand;
 
 mod api;
+mod connect_flight;
 mod connection_registry;
 mod endpoint_state;
 mod lifecycle;
@@ -79,7 +79,6 @@ use connection_registry::ConnectionRegistry;
 use endpoint_state::EndpointLease;
 use endpoint_state::EndpointStateStore;
 use lifecycle::ClientLifecycle;
-use lifecycle::ConnectionCommitFence;
 use nameserver::NameServerHealth;
 
 #[cfg(test)]
@@ -441,310 +440,6 @@ impl<PR: RequestProcessor + Sync + Clone + 'static> TransportClient<PR> {
 }
 
 impl<PR: RequestProcessor + Sync + Clone + 'static> TransportClient<PR> {
-    /// Get existing healthy client or create new connection.
-    ///
-    /// # Flow
-    /// 1. If `addr` is `None` or empty, route to nameserver
-    /// 2. Check the endpoint-session registry for an existing client
-    /// 3. Verify client health (connection.is_healthy() == true)
-    /// 4. If unhealthy or missing, create new connection
-    ///
-    /// # Performance
-    /// - **Fast path**: Single lock acquire + HashMap lookup + health check (< 100ns)
-    /// - **Slow path**: Lock + TCP handshake + TLS (if enabled) (10-50ms)
-    async fn get_and_create_client_until(
-        &self,
-        addr: Option<&CheetahString>,
-        deadline: RequestDeadline,
-    ) -> RocketMQResult<Option<TransportSession<PR>>> {
-        // Route empty addresses to nameserver
-        let target_addr = match addr {
-            None => {
-                return self
-                    .get_and_create_nameserver_client_until(deadline)
-                    .await
-                    .map(|selection| selection.map(|selection| selection.session))
-            }
-            Some(addr) if addr.is_empty() => {
-                return self
-                    .get_and_create_nameserver_client_until(deadline)
-                    .await
-                    .map(|selection| selection.map(|selection| selection.session))
-            }
-            Some(addr) => addr,
-        };
-        deadline.ensure_before_send(target_addr.to_string())?;
-
-        if let Some(client) = self.connection_registry.healthy_session(target_addr, None) {
-            return Ok(Some(client));
-        }
-        self.create_client_until(target_addr, deadline).await
-    }
-
-    /// Create new client connection with double-checked locking pattern.
-    ///
-    /// # Concurrency Strategy
-    ///
-    /// Uses **double-checked locking** to prevent thundering herd:
-    /// 1. **Check 1**: Quick lookup before TCP connect (avoids redundant connects)
-    /// 2. **Release lock**: Perform TCP connect WITHOUT holding lock
-    /// 3. **Check 2**: Re-acquire lock and verify no other task created connection
-    /// 4. **Insert**: Store the new client in the endpoint-session registry
-    ///
-    /// # Performance
-    ///
-    /// **Before (holding lock during connect)**:
-    /// ```text
-    /// Thread 1: [====== LOCK ======][==== CONNECT (50ms) ====][==== INSERT ====]
-    /// Thread 2:                      [waiting.....................][LOCK]
-    /// Thread 3:                      [waiting.....................][LOCK]
-    /// Total: ~50ms * 3 = 150ms wasted
-    /// ```
-    ///
-    /// **After (lock-free connect)**:
-    /// ```text
-    /// Thread 1: [== LOCK ==][RELEASE]→[CONNECT 50ms]→[LOCK][INSERT]
-    /// Thread 2: [== LOCK ==][RELEASE]→[CONNECT 50ms]→[LOCK][cached!]
-    /// Thread 3: [== LOCK ==][RELEASE]→[CONNECT 50ms]→[LOCK][cached!]
-    /// Total: ~50ms + small lock overhead
-    /// ```
-    ///
-    /// # Arguments
-    ///
-    /// * `addr` - Target address (e.g., "127.0.0.1:10911")
-    /// * `duration` - Connection timeout
-    ///
-    /// # Returns
-    ///
-    /// * `Some(client)` - Successfully connected (either new or cached)
-    /// * `None` - Connection failed or timed out (or circuit breaker OPEN)
-    #[cfg(test)]
-    async fn create_client(&self, addr: &CheetahString, duration: Duration) -> Option<TransportSession<PR>> {
-        match self.create_client_until(addr, RequestDeadline::after(duration)).await {
-            Ok(client) => client,
-            Err(error) => {
-                error!(remote_addr = %addr, error = ?error, "Failed to create remoting client");
-                None
-            }
-        }
-    }
-
-    async fn create_client_until(
-        &self,
-        addr: &CheetahString,
-        deadline: RequestDeadline,
-    ) -> RocketMQResult<Option<TransportSession<PR>>> {
-        self.create_client_with_lease_until(addr, None, None, deadline).await
-    }
-
-    async fn create_client_for_nameserver_until(
-        &self,
-        addr: &CheetahString,
-        endpoint: NameServerEndpoint,
-        lease: EndpointLease,
-        deadline: RequestDeadline,
-    ) -> RocketMQResult<Option<TransportSession<PR>>> {
-        self.create_client_with_lease_until(addr, Some(endpoint), Some(lease), deadline)
-            .await
-    }
-
-    async fn create_client_with_lease_until(
-        &self,
-        addr: &CheetahString,
-        configured_nameserver: Option<NameServerEndpoint>,
-        lease: Option<EndpointLease>,
-        deadline: RequestDeadline,
-    ) -> RocketMQResult<Option<TransportSession<PR>>> {
-        deadline.ensure_before_send(addr.to_string())?;
-        if !self.can_commit_endpoint_lease(lease.as_ref()) {
-            return Ok(None);
-        }
-        if let Some(client) = self.connection_registry.healthy_session(addr, lease.as_ref()) {
-            return Ok(Some(client));
-        }
-
-        // Capture the exact worker owner before publishing a connection flight.
-        // A shutdown advances the worker epoch, so a request admitted before a
-        // restart can never spawn or commit through a replacement worker group.
-        let worker_owner = self
-            .capture_worker_task_owner()
-            .ok_or(RocketMQError::ClientNotStarted)?;
-        let (flight, leader) = self.connection_registry.acquire_flight(addr.clone(), lease.clone());
-        if leader {
-            #[cfg(test)]
-            self.wait_for_leader_before_worker_spawn_test_hook().await;
-            let client = self.clone();
-            let flight_for_task = flight.clone();
-            let target = addr.clone();
-            let connect_timeout = self.tokio_client_config.connect.timeout;
-            let target_for_task = target.clone();
-            let lease_for_task = lease.clone();
-            let commit_fence = worker_owner.commit_fence();
-            let spawned = self.spawn_worker_task_with_owner(
-                &worker_owner,
-                format!("rocketmq.transport.connect.{target}"),
-                async move {
-                    client.connect_attempts.fetch_add(1, Ordering::Relaxed);
-                    let result = client
-                        .connect_endpoint_until(
-                            &target_for_task,
-                            configured_nameserver,
-                            lease_for_task,
-                            RequestDeadline::after(connect_timeout),
-                            &commit_fence,
-                        )
-                        .await;
-                    let result = if client.matches_connection_commit_fence(&commit_fence) {
-                        result
-                    } else {
-                        Err(RocketMQError::ClientNotStarted)
-                    };
-                    flight_for_task.complete(result);
-                    client
-                        .connection_registry
-                        .remove_flight_if_matches(&target_for_task, &flight_for_task);
-                },
-            );
-            if spawned.is_none() {
-                flight.complete(Err(RocketMQError::ClientNotStarted));
-                self.connection_registry.remove_flight_if_matches(&target, &flight);
-            }
-        }
-        flight.wait(deadline, addr).await
-    }
-
-    async fn connect_endpoint_until(
-        &self,
-        addr: &CheetahString,
-        configured_nameserver: Option<NameServerEndpoint>,
-        lease: Option<EndpointLease>,
-        deadline: RequestDeadline,
-        commit_fence: &ConnectionCommitFence,
-    ) -> RocketMQResult<Option<TransportSession<PR>>> {
-        deadline.ensure_before_send(addr.to_string())?;
-        if !self.matches_connection_commit_fence(commit_fence) {
-            return Err(RocketMQError::ClientNotStarted);
-        }
-        if !self.can_commit_endpoint_lease(lease.as_ref()) {
-            return Ok(None);
-        }
-        if let Some(client) = self.connection_registry.healthy_session(addr, lease.as_ref()) {
-            return Ok(Some(client));
-        }
-
-        let allowed = match lease.as_ref() {
-            Some(lease) => self
-                .nameserver_health
-                .connection_admission_if_current(addr, lease, || self.endpoint_state.is_current(lease))
-                .is_some_and(|admission| admission != crate::clients::reconnect::CircuitAdmission::Rejected),
-            None => self
-                .direct_circuit_breakers
-                .entry(addr.clone())
-                .or_insert_with(crate::clients::reconnect::CircuitBreaker::default_breaker)
-                .allow_request(),
-        };
-        if !allowed {
-            warn!("Circuit breaker OPEN for {}, rejecting connection attempt", addr);
-            return Ok(None);
-        }
-
-        let session_target = match configured_nameserver.as_ref() {
-            Some(endpoint) => match endpoint.connect_target() {
-                Some(target) => SessionConnectTarget::Resolved(target.clone()),
-                None => SessionConnectTarget::Legacy(endpoint.compatibility_address().to_string()),
-            },
-            None => SessionConnectTarget::Legacy(addr.to_string()),
-        };
-        let transport_config = (*self.tokio_client_config).clone();
-        let frame_limits = self.frame_limits;
-
-        let transport_security = self.transport_security.clone();
-        let connect_result = TransportSession::connect_target_with_service_context_until_and_telemetry(
-            &self.service_context,
-            session_target,
-            self.cmd_handler.clone(),
-            self.tx.as_ref(),
-            transport_config,
-            frame_limits,
-            deadline,
-            self.telemetry.clone(),
-        )
-        .await;
-        let connect_result = match transport_security {
-            Some(transport_security) => connect_result.map(|client| client.with_transport_security(transport_security)),
-            None => connect_result,
-        };
-
-        match connect_result {
-            Ok(new_client) => {
-                #[cfg(test)]
-                self.wait_for_connect_completion_test_hook().await;
-                if !self.matches_connection_commit_fence(commit_fence) {
-                    new_client.begin_drain();
-                    let _ = new_client.close_with_report(Duration::from_secs(1)).await;
-                    return Err(RocketMQError::ClientNotStarted);
-                }
-                if !self.can_commit_endpoint_lease(lease.as_ref()) {
-                    new_client.begin_drain();
-                    let _ = new_client.close_with_report(Duration::from_secs(1)).await;
-                    return Ok(None);
-                }
-                match lease.as_ref() {
-                    Some(lease) => self
-                        .nameserver_health
-                        .record_connection_success_if_current(addr, lease, || self.endpoint_state.is_current(lease)),
-                    None => {
-                        if let Some(mut breaker) = self.direct_circuit_breakers.get_mut(addr) {
-                            breaker.record_success();
-                        }
-                    }
-                }
-                let session_lease = lease.clone();
-                match self
-                    .connection_registry
-                    .insert_session(addr.clone(), new_client.clone(), lease, || {
-                        self.matches_connection_commit_fence(commit_fence)
-                            && self.can_commit_endpoint_lease(session_lease.as_ref())
-                    }) {
-                    Some(client) => {
-                        info!("Successfully created client for {}", addr);
-                        Ok(Some(client))
-                    }
-                    None => {
-                        new_client.begin_drain();
-                        let _ = new_client.close_with_report(Duration::from_secs(1)).await;
-                        if self.matches_connection_commit_fence(commit_fence) {
-                            Ok(None)
-                        } else {
-                            Err(RocketMQError::ClientNotStarted)
-                        }
-                    }
-                }
-            }
-            Err(error) => {
-                error!(remote_addr = %addr, error = ?error, "Failed to connect");
-                if !self.matches_connection_commit_fence(commit_fence) {
-                    return Err(RocketMQError::ClientNotStarted);
-                }
-                match lease.as_ref() {
-                    Some(lease) => self
-                        .nameserver_health
-                        .record_connection_failure_if_current(addr, lease, || self.endpoint_state.is_current(lease)),
-                    None => {
-                        if let Some(mut breaker) = self.direct_circuit_breakers.get_mut(addr) {
-                            breaker.record_failure();
-                        }
-                    }
-                }
-                Err(error)
-            }
-        }
-    }
-
-    fn can_commit_endpoint_lease(&self, lease: Option<&EndpointLease>) -> bool {
-        lease.is_none_or(|lease| self.endpoint_state.is_current(lease))
-    }
-
     /// Scans persistent sessions and removes those that are unhealthy or idle.
     fn scan_idle_connections(&self) {
         let idle_threshold = self.tokio_client_config.maintenance.idle_scan_interval;
